Update VMT contact gagehugo key
This change updates gagehugo's public gpg key, as he has since
learned the valuable lesson in creating a secure backup of his
secret key to avoid the headache of losing it when formatting
a disk.

He was however, responsible enough to have generated a
revocation certificate prior, and the old key has since been
revoked.

Change-Id: I714bf925ec728828db09e7efd86d7190e5e615b7