Refresh public OpenPGP key for Jeremy Stanley
I explicitly maintain a <= 1 year expiration on my public OpenPGP
key and bump it periodically. As a result, the convenience export we
publish should also be refreshed with some regularity so that an old
expired version is not served in error. Update my key here with a
more recent export.

Change-Id: I6914dc0799538fe4d8106593f414e5bfbaf65137