Remove UUID guessing example from C1 report class
Several recent examples have made it clear that some objects in
various services' APIs are keyed with UUIDs which are not
safeguarded or we otherwise don't position this information as
sensitive in obvious ways. We may still consider some reports a C1
if they genuinely hinge on an attacker guessing or socially
engineering an administrator to divulge an arbitrary type 4 UUID,
but this rule is not as solid an indicator of a C1 report as it was
in 2015 when this example was originally added.

Change-Id: Id9a768d3cb3880c4d8f28a45adc924edb8b5dc4a