Make VMT contact list more discoverable
Get rid of the outdated section for the long gone Security Project,
and move the VMT contact info from it to near the top of the main
security.o.o page. Also switch references in the process document to
link that list instead of going to the LP group page (which made
obtaining contact information a challenge).

Change-Id: I6aaf4da8bff51bc63706fc20e9f5f68d6e9b0fe4