Clarify our affected versions conventions
The old impact description template was slightly misleading in its
use of <= to clamp upper bounds of affected versions. For many years
we've actually been using a strict < of the next possible SemVer
patchlevel version, so correct the examples and add a brief
paragraph to explain the construction in greater detail.

Change-Id: I44db2454bd1cd8691f445a0dcd403b8fa2681de3