Import VMT oversight information from governance
The OpenStack TC has decided to stop using its "governance tags"
mechanism for recording specific details about project deliverables.
We previously relied on a vulnerability:managed tag to indicate the
deliverables overseen by the VMT, as well as documenting
expectations for the teams responsible for them.

Tags, being deliverable-specific rather than repository-specific,
were never a great fit for us. When bringing this information into
our own documentation, it's now reworked as a list of specific Git
repositories for simplicity and granularity. The expectations have
also been edited and shortened in order to accommodate this change,
but are still effectively the same as they were in governance.

Change-Id: Ie3c0cc38fc071716420c12b3f6de4a320428bd04