Drop references for the old security blog
Members of the OSSG maintained an OpenStack Security Blog between
2016 and 2017, but it's been abandoned for nearly 5 years now and
none of the currently involved contributors in the SIG have access
to that site nor available time to contribute new articles. Remove
the reference for now, it can always be added back if the blog is
resurrected or replaced in the future.

Change-Id: I04ba8b7cd734707406e480142a6b01df8900f1e9