Add relevent Ironic-governance repos to VMT
This adds the most actively supported projects run in production to the
list of projects under the VMT. This includes all of Ironic's clients,
all of Ironic's networking plugins for neutron, and our redfish client
(sushy).

It's worth noting that some of these items, particularly the
inspector-related ones, will not be here long as they are being
deprecated; but any issues reported there should still be VMT-managed.

Change-Id: I24fd72619965d21651e5e018ccb617f4200c4e30